Refactor: create separate components for each tabs in profile page
